add yolo requirements

+++ yolo/gen-onnx.py
@@ -1,17 +1,19 @@
 import torch
 import os 
 
+if not os.path.exists("/models/ultralytics_yolov5_master/"):
+    print("You should prepare ultralytics yolov5 on /models/, which can be found on /root/.cache/torch/hub/ultralytics_yolov5_master/")
+    exit(1)
 if not os.path.exists("./yolov5m.pt"):
     os.system(
         "cp /models/yolov5m.pt ."
-        #"wget https://github.com/ultralytics/yolov5/releases/download/v6.0/yolov5m.pt"
     )
 else:
     print("Load Local PTH FILE")
-    
-model = torch.hub.load('ultralytics/yolov5', 'yolov5m', pretrained=True)
 
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+model = torch.hub.load('/models/ultralytics_yolov5_master//', 'yolov5m', source="local")
+
+device = torch.device("cpu")
 # print(device)
 random_input = torch.randn(24, 3, 640, 640)
 # # print(tensor.device.type)
